+++ includes/hooks/actions/update_postmeta.php
@@ -17,31 +17,33 @@ function wpnfps_update_postmeta( $meta_id, $object_id, $meta_key, $meta_value ) 
         if ( '_wpforge_network_post_sharing_share_with' === $meta_key ) {
             $share_with = maybe_unserialize( $meta_value );
 
-            if ( $share_with ) {
-                foreach( $share_with as $blog_id ) {
-                    if( $blog_id ) {
-                        $remote_id = wpfnps_update_post_share( $blog_id, $object_id );
+            if ( is_array( $share_with ) ) {
+                if ( $share_with ) {
+                    foreach( $share_with as $blog_id ) {
+                        if( $blog_id ) {
+                            $remote_id = wpfnps_update_post_share( $blog_id, $object_id );
 
-                        if ( $remote_id ) {
-                            update_post_meta( $object_id, "_remote_share_id_{$blog_id}", $remote_id );
+                            if ( $remote_id ) {
+                                update_post_meta( $object_id, "_remote_share_id_{$blog_id}", $remote_id );
+                            }
                         }
                     }
-                }
-            }
 
-            $delete_from = array_diff( $current_shares, $share_with );
+                    $delete_from = array_diff( $current_shares, $share_with );
 
-            foreach( $delete_from as  $maybe_delete ) {
-                if ( $maybe_delete && ! in_array( $maybe_delete, $share_with ) ) {
-                    wpfnps_delete_remote_share( $maybe_delete, $object_id );
-                    delete_post_meta( $object_id, "_remote_share_id_{$maybe_delete}" );
+                    foreach( $delete_from as  $maybe_delete ) {
+                        if ( $maybe_delete && ! in_array( $maybe_delete, $share_with ) ) {
+                            wpfnps_delete_remote_share( $maybe_delete, $object_id );
+                            delete_post_meta( $object_id, "_remote_share_id_{$maybe_delete}" );
+                        }
+                    }
                 }
             }
 
             break;
         }
 
-        if ( $current_shares ) {
+        if ( is_array( $current_shares )  && $current_shares ){
             remove_action( 'update_postmeta', 'wpnfps_update_postmeta' );
             foreach( $current_shares as $site_id ) {
                 $remote_id = wpfnps_remote_share_id( $object_id, $site_id );
@@ -93,8 +95,10 @@ function wpfnps_delete_remote_images( $post_id ) {
 
     $gallery_images = get_field( 'image_gallery', $post_id );
 
-    foreach( $gallery_images as $image ) {
-        wp_delete_post( $image['ID'] );
+    if ( is_array( $gallery_images) && $gallery_images ) {
+        foreach( $gallery_images as $image ) {
+            wp_delete_post( $image['ID'] );
+        }
     }
 
 }
